requirements.txt
crewai
crewai-tools
tavily-python
langchain-google-genai
langchain-tavily
streamlit
python-dotenv
langchain
langchain-groq
langchain-community
langgraph
arxiv
plotly
markdown    
xhtml2pdf   